Move MandatoryFieldChecker to their own module
