+++ src/eblint/checkers/__init__.py
@@ -1,2 +1,3 @@
 from .base_checker import Checker
 from .violation import Violation
+from .mandatory_field_checker import MandatoryFieldChecker
